move ColumnQueryMapper into decorateColumnMethods.ts
