+++ src/charts/BaseChart.ts
@@ -10,9 +10,9 @@ import type { BaseDialog } from "@/utilities/Dialog";
 import type { DataColumn, FieldName, GuiSpecs, Quantiles } from "./charts";
 import type Dimension from "@/datastore/Dimension";
 import { g } from "@/lib/utils";
-import { serialiseChart, initialiseChartConfig, ColumnQueryMapper } from "./chartConfigUtils";
+import { serialiseChart, initialiseChartConfig } from "./chartConfigUtils";
 import type { MultiColumnQuery } from "@/links/link_utils";
-import { decorateChartColumnMethods, loadColumnData } from "@/datastore/decorateColumnMethod";
+import { ColumnQueryMapper, decorateChartColumnMethods, loadColumnData } from "@/datastore/decorateColumnMethod";
 import type { FieldSpec, FieldSpecs } from "@/lib/columnTypeHelpers";
 import getParamsGuiSpec from "./dialogs/utils/ParamsSettingGui";
 export type ChartEventType = string;
